Controlador (enlace entre presentacion - retorfit) incorporado pero incompleto)

+++ app/src/main/java/curso/alberto/sunriseweatherapp/data/WeatherApi/Main.kt
@@ -5,7 +5,7 @@ package curso.alberto.sunriseweatherapp.data.WeatherApi
 
 data class Main(
 
-    val feelsLike: Double,
+    val feelsLike: Int,
 
     val humidity: Int,
